Reorganized directory structure.
