--- src/testing/volttron/__init__.py
@@ -1,9 +1,0 @@
-from testing.volttron.client_mock import TestClient
-from testing.volttron.server_mock import TestServer
-#from testing.volttron.platformwrapper import PlatformWrapper
-
-__all__ = [
-    "TestClient",
-    "TestServer",
-#    "PlatformWrapper"
-]
