+++ src/volttrontesting/__init__.py
@@ -1,0 +1,9 @@
+from volttrontesting.client_mock import TestClient
+
+#from volttrontesting.volttron.platformwrapper import PlatformWrapper
+
+__all__ = [
+    "TestClient",
+    "TestServer",
+#    "PlatformWrapper"
+]
