+++ src/volttrontesting/utils.py
@@ -182,8 +182,8 @@ def validate_published_device_data(expected_headers, expected_message,
 class AgentMock:
     """
     The purpose for this parent class is to be used for unit
-    testing of agents. It takes in the class methods of other
-    classes, turns them into it's own mock methods. For testing,
+    volttrontesting of agents. It takes in the class methods of other
+    classes, turns them into it's own mock methods. For volttrontesting,
     dynamically replace the agent's current base class with this
     class, while passing in the agent's current classes as arguments.
